+++ src/main/java/com/aladdinsys/api/exercise/repository/ExerciseRepository.java
@@ -1,6 +1,6 @@
-package com.aladdinsys.api.exercise_resource.repository;
+package com.aladdinsys.api.exercise.repository;
 
-import com.aladdinsys.api.exercise_resource.entity.ExerciseEntity;
+import com.aladdinsys.api.exercise.entity.ExerciseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExerciseRepository extends JpaRepository<ExerciseEntity, Long> {
